// src/interfaces/gossip.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.6
//   protoc               v7.34.0
// source: gossip.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from '@nestjs/microservices';
import { Observable } from 'rxjs';
import { Empty, UUID } from './shared';

export const protobufPackage = 'event_store.client.gossip';

export interface ClusterInfo {
  members: MemberInfo[];
}

export interface EndPoint {
  address: string;
  port: number;
}

export interface MemberInfo {
  instanceId: UUID | undefined;
  timeStamp: number;
  state: MemberInfo_VNodeState;
  isAlive: boolean;
  httpEndPoint: EndPoint | undefined;
}

export enum MemberInfo_VNodeState {
  Initializing = 0,
  DiscoverLeader = 1,
  Unknown = 2,
  PreReplica = 3,
  CatchingUp = 4,
  Clone = 5,
  Follower = 6,
  PreLeader = 7,
  Leader = 8,
  Manager = 9,
  ShuttingDown = 10,
  Shutdown = 11,
  ReadOnlyLeaderless = 12,
  PreReadOnlyReplica = 13,
  ReadOnlyReplica = 14,
  ResigningLeader = 15,
  UNRECOGNIZED = -1,
}

export const EVENT_STORE_CLIENT_GOSSIP_PACKAGE_NAME =
  'event_store.client.gossip';

export interface GossipClient {
  read(request: Empty): Observable<ClusterInfo>;
}

export interface GossipController {
  read(
    request: Empty,
  ): Promise<ClusterInfo> | Observable<ClusterInfo> | ClusterInfo;
}

export function GossipControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ['read'];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(
        constructor.prototype,
        method,
      );
      GrpcMethod('Gossip', method)(
        constructor.prototype[method],
        method,
        descriptor,
      );
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(
        constructor.prototype,
        method,
      );
      GrpcStreamMethod('Gossip', method)(
        constructor.prototype[method],
        method,
        descriptor,
      );
    }
  };
}

export const GOSSIP_SERVICE_NAME = 'Gossip';