// src/interfaces/operations.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.6
//   protoc               v7.34.0
// source: operations.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from '@nestjs/microservices';
import { Observable } from 'rxjs';
import { Empty } from './shared';

export const protobufPackage = 'event_store.client.operations';

export interface StartScavengeReq {
  options: StartScavengeReq_Options | undefined;
}

export interface StartScavengeReq_Options {
  threadCount: number;
  startFromChunk: number;
}

export interface StopScavengeReq {
  options: StopScavengeReq_Options | undefined;
}

export interface StopScavengeReq_Options {
  scavengeId: string;
}

export interface ScavengeResp {
  scavengeId: string;
  scavengeResult: ScavengeResp_ScavengeResult;
}

export enum ScavengeResp_ScavengeResult {
  Started = 0,
  InProgress = 1,
  Stopped = 2,
  UNRECOGNIZED = -1,
}

export interface SetNodePriorityReq {
  priority: number;
}

export const EVENT_STORE_CLIENT_OPERATIONS_PACKAGE_NAME =
  'event_store.client.operations';

export interface OperationsClient {
  startScavenge(request: StartScavengeReq): Observable<ScavengeResp>;

  stopScavenge(request: StopScavengeReq): Observable<ScavengeResp>;

  shutdown(request: Empty): Observable<Empty>;

  mergeIndexes(request: Empty): Observable<Empty>;

  resignNode(request: Empty): Observable<Empty>;

  setNodePriority(request: SetNodePriorityReq): Observable<Empty>;

  restartPersistentSubscriptions(request: Empty): Observable<Empty>;
}

export interface OperationsController {
  startScavenge(
    request: StartScavengeReq,
  ): Promise<ScavengeResp> | Observable<ScavengeResp> | ScavengeResp;

  stopScavenge(
    request: StopScavengeReq,
  ): Promise<ScavengeResp> | Observable<ScavengeResp> | ScavengeResp;

  shutdown(request: Empty): Promise<Empty> | Observable<Empty> | Empty;

  mergeIndexes(request: Empty): Promise<Empty> | Observable<Empty> | Empty;

  resignNode(request: Empty): Promise<Empty> | Observable<Empty> | Empty;

  setNodePriority(
    request: SetNodePriorityReq,
  ): Promise<Empty> | Observable<Empty> | Empty;

  restartPersistentSubscriptions(
    request: Empty,
  ): Promise<Empty> | Observable<Empty> | Empty;
}

export function OperationsControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = [
      'startScavenge',
      'stopScavenge',
      'shutdown',
      'mergeIndexes',
      'resignNode',
      'setNodePriority',
      'restartPersistentSubscriptions',
    ];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(
        constructor.prototype,
        method,
      );
      GrpcMethod('Operations', method)(
        constructor.prototype[method],
        method,
        descriptor,
      );
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(
        constructor.prototype,
        method,
      );
      GrpcStreamMethod('Operations', method)(
        constructor.prototype[method],
        method,
        descriptor,
      );
    }
  };
}

export const OPERATIONS_SERVICE_NAME = 'Operations';