// src/interfaces/persistent.ts
export interface ReadResp {
  event?: ReadResp_ReadEvent | undefined;
  subscriptionConfirmation?: ReadResp_SubscriptionConfirmation | undefined;
}

export interface ReadResp_ReadEvent {
  event: ReadResp_ReadEvent_RecordedEvent | undefined;
  link: ReadResp_ReadEvent_RecordedEvent | undefined;
  commitPosition?: number | undefined;
  noPosition?: Empty | undefined;
  retryCount?: number | undefined;
  noRetryCount?: Empty | undefined;
}

export interface ReadResp_ReadEvent_RecordedEvent {
  id: UUID | undefined;
  streamIdentifier: StreamIdentifier | undefined;
  streamRevision: number;
  preparePosition: number;
  commitPosition: number;
  metadata: { [key: string]: string };
  customMetadata: Uint8Array;
  data: Uint8Array;
}

export interface ReadResp_ReadEvent_RecordedEvent_MetadataEntry {
  key: string;
  value: string;
}

export interface ReadResp_SubscriptionConfirmation {
  subscriptionId: string;
}
// ...
export interface GetInfoReq {
  options: GetInfoReq_Options | undefined;
}

export interface GetInfoReq_Options {
  streamIdentifier?: StreamIdentifier | undefined;
  all?: Empty | undefined;
  groupName: string;
}

export interface GetInfoResp {
  subscriptionInfo: SubscriptionInfo | undefined;
}

export interface SubscriptionInfo {
  eventSource: string;
  groupName: string;
  status: string;
  connections: SubscriptionInfo_ConnectionInfo[];
  averagePerSecond: number;
  totalItems: number;
  countSinceLastMeasurement: number;
  lastCheckpointedEventPosition: string;
  lastKnownEventPosition: string;
  resolveLinkTos: boolean;
  startFrom: string;
  messageTimeoutMilliseconds: number;
  extraStatistics: boolean;
  maxRetryCount: number;
  liveBufferSize: number;
  bufferSize: number;
  readBatchSize: number;
  checkPointAfterMilliseconds: number;
  minCheckPointCount: number;
  maxCheckPointCount: number;
  readBufferCount: number;
  liveBufferCount: number;
  retryBufferCount: number;
  totalInFlightMessages: number;
  outstandingMessagesCount: number;
  namedConsumerStrategy: string;
  maxSubscriberCount: number;
  parkedMessageCount: number;
}

export interface SubscriptionInfo_ConnectionInfo {
  from: string;
  username: string;
  averageItemsPerSecond: number;
  totalItems: number;
  countSinceLastMeasurement: number;
  observedMeasurements: SubscriptionInfo_Measurement[];
  availableSlots: number;
  inFlightMessages: number;
  connectionName: string;
}

export interface SubscriptionInfo_Measurement {
  key: string;
  value: number;
}
// ...
export interface ListResp {
  subscriptions: SubscriptionInfo[];
}