// src/interfaces/shared.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.7.3
//   protoc               v3.20.3
// source: shared.proto

/* eslint-disable */
import { Empty as Empty1 } from "./google/protobuf/empty";

export const protobufPackage = "event_store.client";

export interface UUID {
  structured?: UUID_Structured | undefined;
  string?: string | undefined;
}

export interface UUID_Structured {
  mostSignificantBits: number;
  leastSignificantBits: number;
}

export interface Empty {
}

export interface StreamIdentifier {
  streamName: Uint8Array;
}

/** @deprecated */
export interface AllStreamPosition {
  commitPosition: number;
  preparePosition: number;
}

export interface WrongExpectedVersion {
  currentStreamRevision?: number | undefined;
  currentNoStream?: Empty1 | undefined;
  expectedStreamPosition?: number | undefined;
  expectedAny?: Empty1 | undefined;
  expectedStreamExists?: Empty1 | undefined;
  expectedNoStream?: Empty1 | undefined;
}

export interface AccessDenied {
}

export interface StreamDeleted {
  streamIdentifier: StreamIdentifier | undefined;
}

export interface Timeout {
}

export interface Unknown {
}

export interface InvalidTransaction {
}

export interface MaximumAppendSizeExceeded {
  maxAppendSize: number;
}

/** @deprecated */
export interface BadRequest {
  message: string;
}

export const EVENT_STORE_CLIENT_PACKAGE_NAME = "event_store.client";