// src/interfaces/monitoring.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.6
//   protoc               v7.34.0
// source: monitoring.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from '@nestjs/microservices';
import { Observable } from 'rxjs';

export const protobufPackage = 'event_store.client.monitoring';

export interface StatsReq {
  useMetadata: boolean;
  refreshTimePeriodInMs: number;
}

export interface StatsResp {
  stats: { [key: string]: string };
}

export interface StatsResp_StatsEntry {
  key: string;
  value: string;
}

export const EVENT_STORE_CLIENT_MONITORING_PACKAGE_NAME =
  'event_store.client.monitoring';

export interface MonitoringClient {
  stats(request: StatsReq): Observable<StatsResp>;
}

export interface MonitoringController {
  stats(request: StatsReq): Observable<StatsResp>;
}

export function MonitoringControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ['stats'];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(
        constructor.prototype,
        method,
      );
      GrpcMethod('Monitoring', method)(
        constructor.prototype[method],
        method,
        descriptor,
      );
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(
        constructor.prototype,
        method,
      );
      GrpcStreamMethod('Monitoring', method)(
        constructor.prototype[method],
        method,
        descriptor,
      );
    }
  };
}

export const MONITORING_SERVICE_NAME = 'Monitoring';