// src/interfaces/users.ts
export interface DetailsReq {
  options: DetailsReq_Options | undefined;
}

export interface DetailsReq_Options {
  loginName: string;
}

export interface DetailsResp {
  userDetails: DetailsResp_UserDetails | undefined;
}

export interface DetailsResp_UserDetails {
  loginName: string;
  fullName: string;
  groups: string[];
  lastUpdated: DetailsResp_UserDetails_DateTime | undefined;
  disabled: boolean;
}

export interface DetailsResp_UserDetails_DateTime {
  ticksSinceEpoch: number;
}